requirements.txt
Django~=4.0.2
django-markdownx>=4.0.0b1
django-colorfield
fontawesomefree==6.1.1
django-allauth==0.51.0
django-feather==0.3.0
django-recurrence==1.11.1
django-q==1.3.9